comments plus smite some more
